\docType{data}
\name{CD4}
\alias{CD4}
\title{The CD4 dataset}
\format{data frame with 6 variables and 1817 observations.}
\usage{
CD4
}
\description{
This dataset is a subset of the Multicenter AIDS cohort study. The data
contain repeated measurements of physical examinations and CD4 percentages
of 283 homosexual men who became HIV-positive between year 1984 and 1991.
}
\details{
It contains the following variables:
\describe{
\item{subj}{the subject (man) indicator }
\item{Time}{the time in years}
\item{Smooking}{the smoking status}
\item{Age}{age at HIV infection}
\item{PreCD4}{the pre-infection CD4 percentage}
\item{CD4}{the CD4 percentage}
}
}
\keyword{datasets}
