\docType{data}
\name{wages}
\alias{wages}
\title{The wages dataset}
\format{a \code{data.frame} with 7 variables and 6402 observations.}
\usage{
wages
}
\description{
This dataset is taken from the National Longitudinal Survey of Youth, collected in U.S.A. The data contain repeated measurements for 888 individuals of age 14 -- 17 years old.
}
\details{
It contains the following variables:
\describe{
\item{id}{the subject (individual) indicator }
\item{exper}{the duration of work experience in years}
\item{hgc}{the highest grade completed by the indiviadual}
\item{race}{the race of the individual: black, hispanic and white}
\item{resp}{the hourly wage}
\item{r1}{the dummy variable for black individuals}
\item{r2}{the dummy variable for hispanic individuals}
}
}
\references{
Murnane, R. J., Willett, J. B., and Boudett, K. P. (1999). Do male dropouts benefit
from obtaining a GED, postsecondary education, and training? \emph{Evaluation Review},
23(5):475-503.
}
\keyword{datasets}